requirements.txt
flask==2.3.2
python-dotenv==1.0.0
zxcvbn==4.4.28
requests==2.31.0
openai>=1.0.0
scikit-learn==1.3.0
joblib==1.3.2 